refactor(prisma-imports): modify prisma imports used by modules to account for refactor
