+++ apps/ourvoice-api/src/modules/comment/comment.module.ts
@@ -2,7 +2,7 @@ import { CommentResolver } from './comment.resolver';
 import { CommentService } from './comment.service';
 import { CommentRepository } from './comment.repository';
 import { Module } from '@nestjs/common';
-import { PrismaModule } from 'src/database/prisma.module';
+import { PrismaModule } from 'src/database/main/prisma.module';
 
 @Module({
   imports: [PrismaModule],
